Added in changes required to upgrade main sites app to Django 2.2.
In links/forms.py, a queryset was accessed which hit the database at
interpretation time, before the necessary CMS tables could be created.
This was moved into the try block to work around this error.

In migration 0024, the accessing of the icon_font from the DB causes the
transaction block to fail. Since this migration doesn't actually modify
anything in the our specific case, the quickest way that didn't involve
excessive DB debugging was to comment out this line.